adding CartService.java and fixed on model have type data BigDecimal but ont dtos i type integer so errro while i create logic for getCart method lol

+++ backend/src/main/java/com/ecommerce/backend/dtos/CartItemResponse.java
@@ -2,6 +2,7 @@ package com.ecommerce.backend.dtos;
 
 import lombok.Builder;
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -9,7 +10,7 @@ public class CartItemResponse {
     private Long id; // ID dari CartItem ini
     private Long productId;
     private String productName;
-    private Integer price;
+    private BigDecimal price;
     private Integer quantity;
-    private Integer subTotal; // Hasil kali price * quantity
+    private BigDecimal subTotal; // Hasil kali price * quantity
 }
