+++ backend/src/main/java/com/ecommerce/backend/dtos/CartResponse.java
@@ -3,6 +3,7 @@ package com.ecommerce.backend.dtos;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -10,5 +11,5 @@ import java.util.List;
 public class CartResponse {
     private Long cartId;
     private List<CartItemResponse> items;
-    private Integer totalPrice; // Total harga seluruh barang di keranjang
+    private BigDecimal totalPrice; // Total harga seluruh barang di keranjang
 }
